- splits home fragment to dedicated composables. ForYouContentTab.kt, CommunityContentTab.kt, HomeScreen.kt and shared FeedComponents.kt.
